requirements.txt
numpy==2.0.1
pandas==2.2.3
geopandas==1.0.1
shapely==2.0.7
scipy==1.10.1
scikit-learn==1.2.0
h3pandas==0.2.6
polygeohasher==1.0.0
python-geohash==0.8.5
tqdm==4.65.0
torch==2.0.1
transformers==4.29.2